+++ internal/impl/snowflake/streaming/rest.go
@@ -173,12 +173,12 @@ type (
 		// current hex-encoded max value, truncated down to 32 bytes
 		MinStrValue *string `json:"minStrValue"`
 		// current hex-encoded max value, truncated up to 32 bytes
-		MaxStrValue  *string    `json:"maxStrValue"`
-		MinIntValue  int128.Num `json:"minIntValue"`
-		MaxIntValue  int128.Num `json:"maxIntValue"`
-		MinRealValue float64    `json:"minRealValue"`
-		MaxRealValue float64    `json:"maxRealValue"`
-		NullCount    int64      `json:"nullCount"`
+		MaxStrValue  *string         `json:"maxStrValue"`
+		MinIntValue  int128.Num      `json:"minIntValue"`
+		MaxIntValue  int128.Num      `json:"maxIntValue"`
+		MinRealValue json.RawMessage `json:"minRealValue"`
+		MaxRealValue json.RawMessage `json:"maxRealValue"`
+		NullCount    int64           `json:"nullCount"`
 		// Currently not tracked
 		DistinctValues int64 `json:"distinctValues"`
 		MaxLength      int64 `json:"maxLength"`